requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.25.0
pytest>=7.4.0
pytest-mock>=3.11.0
pyyaml>=6.0
pathspec>=0.11.0
jinja2>=3.1.0
